requirements.txt
flask
flask_mysqldb
pyyaml